misc/HtmlTestSuite/results/test-0173.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSS Test: ::slotted() should not match via the matches() API, since it's in the wrong scope</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<link href="mailto:emilio@crisal.io" rel="author" title="Emilio Cobos Ãlvarez"/>
<link href="https://drafts.csswg.org/css-scoping/#slotted-pseudo" rel="help"/>
<link href="https://dom.spec.whatwg.org/#dom-element-matches" rel="help"/>
<link href="https://bugzil.la/1544242" rel="help"/>
</head><body><div id="host"><div id="slotted"/></div>
<script>
test(function() {
  let slotted = document.getElementById("slotted");
  host.attachShadow({ mode: "open" }).innerHTML = `&lt;slot&gt;&lt;/slot&gt;`;
  assert_false(slotted.matches("::slotted(div)"), "Shouldn't match ::slotted from the outer tree")
}, "::slotted() doesn't reveal the presence of shadow DOM via matches()");
</script>
</body></html>
